refactor: Move players to new file
